// lib/app/src/agent/plugin/codex/server/api/ServerRequest.ts
// GENERATED CODE! DO NOT MODIFY BY HAND!

// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ApplyPatchApprovalParams } from "./ApplyPatchApprovalParams";
import type { ExecCommandApprovalParams } from "./ExecCommandApprovalParams";
import type { RequestId } from "./RequestId";
import type { ChatgptAuthTokensRefreshParams } from "./v2/ChatgptAuthTokensRefreshParams";
import type { CommandExecutionRequestApprovalParams } from "./v2/CommandExecutionRequestApprovalParams";
import type { DynamicToolCallParams } from "./v2/DynamicToolCallParams";
import type { FileChangeRequestApprovalParams } from "./v2/FileChangeRequestApprovalParams";
import type { McpServerElicitationRequestParams } from "./v2/McpServerElicitationRequestParams";
import type { ToolRequestUserInputParams } from "./v2/ToolRequestUserInputParams";

/**
 * Request initiated from the server and sent to the client.
 */
export type ServerRequest =
  | {
      method: "item/commandExecution/requestApproval";
      id: RequestId;
      params: CommandExecutionRequestApprovalParams;
    }
  | {
      method: "item/fileChange/requestApproval";
      id: RequestId;
      params: FileChangeRequestApprovalParams;
    }
  | { method: "item/tool/requestUserInput"; id: RequestId; params: ToolRequestUserInputParams }
  | {
      method: "mcpServer/elicitation/request";
      id: RequestId;
      params: McpServerElicitationRequestParams;
    }
  | { method: "item/tool/call"; id: RequestId; params: DynamicToolCallParams }
  | {
      method: "account/chatgptAuthTokens/refresh";
      id: RequestId;
      params: ChatgptAuthTokensRefreshParams;
    }
  | { method: "applyPatchApproval"; id: RequestId; params: ApplyPatchApprovalParams }
  | { method: "execCommandApproval"; id: RequestId; params: ExecCommandApprovalParams };